kv, db: move store package to db package
