+++ db/db_test.go
@@ -1,0 +1,1 @@
+package db_test
