--- store/store_test.go
@@ -1,1 +1,0 @@
-package store_test
